+++ .github/workflows/release.yml
@@ -154,13 +154,13 @@ jobs:
             ghcr.io/${{ needs.preconditions.outputs.org_name }}/${{ needs.preconditions.outputs.repo_name }}:${{ needs.check-version.outputs.version }}
             ghcr.io/${{ needs.preconditions.outputs.org_name }}/${{ needs.preconditions.outputs.repo_name }}:latest
           labels: |
-            org.opencontainers.image.title=${{ needs.preconditions.outputs.repo_name }}
-            org.opencontainers.image.description=${{ github.event.repository.description }}
-            org.opencontainers.image.source=${{ github.event.repository.html_url }}
-            org.opencontainers.image.url=${{ github.event.repository.html_url }}
-            org.opencontainers.image.revision=${{ github.sha }}
-            org.opencontainers.image.version=${{ needs.check-version.outputs.version }}
-            org.opencontainers.image.created=${{ needs.check-version.outputs.build_date }}
+            org.opens.image.title=${{ needs.preconditions.outputs.repo_name }}
+            org.opens.image.description=${{ github.event.repository.description }}
+            org.opens.image.source=${{ github.event.repository.html_url }}
+            org.opens.image.url=${{ github.event.repository.html_url }}
+            org.opens.image.revision=${{ github.sha }}
+            org.opens.image.version=${{ needs.check-version.outputs.version }}
+            org.opens.image.created=${{ needs.check-version.outputs.build_date }}
 
       # Build github release
       - name: Delete latest release
