+++ README.md
@@ -1,0 +1,15 @@
+# Morello UI
+A GUI for executing binaries on morello board
+
+## Integration E2E testing
+This repository uses Cypress for testing UI. Specs can be found in `cypress/integration`.
+**Test Types**
+- `smoke` - checks if renders root component
+
+## Environment variable
+We are using dotenv and webpack to mount all env variables to `process.env`
+
+.... write about some
+
+
+
