+++ src/components/Common.js
@@ -44,19 +44,11 @@ export const Wrapper = styled.section(
   })
 )
 
-export const Item = styled.div(({ size = 1, ...props }) => ({
+// Size gooes from 1 - 10 multiplied by 10
+export const Container = styled.div(({ size = 1, styles = {} }) => ({
   ...params.flex,
-  alignItems: 'center',
   width: `${size * 10}%`,
-  padding: 2,
-  ...props,
-}))
-
-export const Container = styled.div(({ size = 10, ...props }) => ({
-  ...params.flex,
-  flexFlow: 'row wrap',
-  width: `${size * 10}%`,
-  ...props,
+  ...styles,
 }))
 
 export const Spacer = styled.div(({ size = 1 }) => ({
