+++ src/components/Header.js
@@ -3,16 +3,20 @@ import React from 'react'
 import { Container } from './Common'
 import dsbdLogo from '../assets/images/logo.png'
 
-export default function Header({ custom }) {
+export default function Header({ children, styles = {} }) {
   const logo = <img src={dsbdLogo} width={'120px'} height={'35px'} />
 
   return (
     <Container
-      height={'80px'}
-      padding={'10px 20px'}
-      justifyContent={'flex-start'}
+      styles={{
+        height: '80px',
+        size: 10,
+        padding: '10px 20px',
+        justifyContent: 'flex-start',
+        ...styles,
+      }}
     >
-      {custom || logo}
+      {children || logo}
     </Container>
   )
 }
